requirements.txt
transformers
torch
torchvision
torchaudio
pandas
scikit-learn
datasets
fastapi
uvicorn[standard]
python-multipart
PyMuPDF
Pillow
